Update README and add console screenshot
Enhanced the `README.md` with a new HTML-styled header section to introduce the FlowSynx Console, improving its visual presentation and structure. Added a reference to a screenshot (`/img/console.jpg`) to provide a visual aid.

Included a new binary file (`console.png`) to represent the FlowSynx Console visually. Refined the textual description in the `README.md` to clarify the console's purpose, target audience, and capabilities.

#89